+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Removed references to `fiveNode` in testsuite (because these networks are no longer five-node, and the network size is less important than whether it's staking or not)
 * Removed `fiveStakingNodeGetValidatorsTest`, which has been superseded by the RPC workflow test
 * Provided `--list` flag to the initializer CLI to simply list the tests registered in the suite
+* Switch to using the `Info.isBootstrapped` endpoint to check for node liveness
 
 # 0.3.1
 * Specify --http-host CLI flag in GetStartCommand to have RPC calls bind to publicIP